Add files via upload
This is a project to find the YouTube video Id from a YouTube Video link. Go through it and understand it correctly.

You and your friends like to share YouTube video links. You must want to keep track of all the videos you watch in your own personal text editor, but you find that keeping the entire link is unnecessary. Then this program is just for all those types of people.

There are two types of youtube videos Url. One looks like this: "https://www.youtube.com/watch?v=RRW2aUSw5vU". This type of Url is generally seen in the Desktop browser. 

And in the mobile version, the youtube video URL looks like "https://youtu.be/KMBBjzp5hdc". So we can get these two types of youtube video URLs as our input. 

To find the Video Id from these URLs. We must what the video really is? So for example: In this URL "https://www.youtube.com/watch?v=RRW2aUSw5vU" the colored part is our Video Id that we want to print as our Output. 

So For this type of URL, "https://www.youtube.com/watch?v=RRW2aUSw5vU" Notice that after the "=" sign, our video id is there. So we can use it as our advantage. And simply print this Video ID.

But in another case, we don't have the = sign. So in this type of URL, "https://youtu.be/KMBBjzp5hdc." there is "/" and after that the video Id is present. First, try to solve this problem with yourself with the above hints. If you not to solve then go through with this program.